rename float_precision_translator to floats
